Move policies to pkg folder
We need to import AdmissionPolicy, ClusterAdmissionPolicy and Policy in the audit-scanner. In order to do that we need to move the code to the pkg folder
