+++ .golangci.yml
@@ -26,9 +26,9 @@ issues:
     - path: internal/pkg/admission/mutating-webhook.go
       linters:
         - dupl
-    - path: apis/policies/v1alpha2/clusteradmissionpolicy_types.go
+    - path: pkg/apis/policies/v1alpha2/clusteradmissionpolicy_types.go
       text: "got 'TypeMeta' want 'typeMeta'"
-    - path: apis/policies/v1alpha2/policyserver_types.go
+    - path: pkg/apis/policies/v1alpha2/policyserver_types.go
       text: "got 'TypeMeta' want 'typeMeta'"
 
 linters:
